move matrix_bit_code from fluid to phi
