+++ paddle/phi/kernels/funcs/CMakeLists.txt
@@ -21,6 +21,7 @@ math_library(im2col)
 math_library(vol2col)
 math_library(softmax DEPS math_function)
 math_library(maxouting)
+math_library(matrix_bit_code)
 
 cc_library(
   phi_data_layout_transform
